Separate repository_util into Galaxy and Toolshed variants.
